Fixup for https://phabricator.services.mozilla.com/D246780
